+++ sklearn/tree/_tree.pyx
@@ -203,7 +203,7 @@ cdef class Tree:
 
     def __cinit__(self, int n_features, object n_classes, int n_outputs,
                  Criterion criterion, double max_depth, int min_samples_split,
-                 int min_samples_leaf, double min_density, int max_features,
+                 int min_samples_leaf, int max_features,
                  int find_split_algorithm, object random_state, int capacity=3):
         """Constructor."""
         # Input/Output layout
@@ -224,7 +224,6 @@ cdef class Tree:
         self.max_depth = max_depth
         self.min_samples_split = min_samples_split
         self.min_samples_leaf = min_samples_leaf
-        self.min_density = min_density
         self.max_features = max_features
         self.find_split_algorithm = find_split_algorithm
         self.random_state = random_state
@@ -270,7 +269,6 @@ cdef class Tree:
                        self.max_depth,
                        self.min_samples_split,
                        self.min_samples_leaf,
-                       self.min_density,
                        self.max_features,
                        self.find_split_algorithm,
                        self.random_state), self.__getstate__())
@@ -338,8 +336,7 @@ cdef class Tree:
         if capacity < self.node_count:
             self.node_count = capacity
 
-    cpdef build(self, np.ndarray X, np.ndarray y,
-                np.ndarray sample_mask=None, np.ndarray X_argsorted=None):
+    cpdef build(self, np.ndarray X, np.ndarray y):
         """Build a decision tree from the training set (X, y).
 
         Parameters
@@ -357,13 +354,6 @@ cdef class Tree:
         if y.dtype != DOUBLE or not y.flags.contiguous:
             y = np.asarray(y, dtype=DOUBLE, order="C")
 
-        if sample_mask is None:
-            sample_mask = np.ones((X.shape[0],), dtype=np.bool)
-
-        if X_argsorted is None:
-            X_argsorted = np.asfortranarray(
-                np.argsort(X.T, axis=1).astype(np.int32).T)
-
         # Pre-allocate some space
         cdef int init_capacity
 
@@ -375,8 +365,11 @@ cdef class Tree:
         self.resize(init_capacity)
         cdef double* buffer_value = <double*> malloc(self.value_stride * sizeof(double))
 
+        n_samples = X.shape[0]
+        sample_indices = np.arange(n_samples, dtype=np.int)
+
         # Build the tree by recursive partitioning
-        self.recursive_partition(X, X_argsorted, y, sample_mask, np.sum(sample_mask), 0, -1, False, buffer_value)
+        self.recursive_partition(X, y, sample_indices, n_samples, 0, -1, False, buffer_value)
 
         # Compactify
         self.resize(self.node_count)
@@ -384,10 +377,9 @@ cdef class Tree:
 
     cdef void recursive_partition(self,
                                   np.ndarray[DTYPE_t, ndim=2, mode="fortran"] X,
-                                  np.ndarray[np.int32_t, ndim=2, mode="fortran"] X_argsorted,
                                   np.ndarray[DOUBLE_t, ndim=2, mode="c"] y,
-                                  np.ndarray sample_mask,
-                                  int n_node_samples,
+                                  np.ndarray[np.int_t, ndim=1, mode="c"] sample_indices,
+                                  int n_samples,
                                   int depth,
                                   int parent,
                                   int is_left_child,
@@ -396,104 +388,93 @@ cdef class Tree:
         # Variables
         cdef Criterion criterion = self.criterion
 
-        cdef DTYPE_t* X_ptr = <DTYPE_t*> X.data
-        cdef int* X_argsorted_ptr = <int*> X_argsorted.data
-        cdef DOUBLE_t* y_ptr = <DOUBLE_t*> y.data
-        cdef BOOL_t* sample_mask_ptr = <BOOL_t*> sample_mask.data
+        cdef DOUBLE_t* y_ptr
+        cdef int y_stride
 
-        cdef int X_stride = <int> X.strides[1] / <int> X.strides[0]
-        cdef int X_argsorted_stride = <int> X_argsorted.strides[1] / <int> X_argsorted.strides[0]
-        cdef int y_stride = <int> y.strides[0] / <int> y.strides[1]
-
-        cdef int n_total_samples = y.shape[0]
         cdef int feature
         cdef double threshold
         cdef double best_error
         cdef double init_error
+        cdef int n_sample_left
+        cdef int n_samples_right
+
+        cdef np.ndarray X_current = None
+        cdef np.ndarray y_current = None
+        cdef np.ndarray sample_indices_left = None
+        cdef np.ndarray sample_indices_right = None
 
         cdef int i
-        cdef np.ndarray sample_mask_left
-        cdef np.ndarray sample_mask_right
-        cdef int n_node_samples_left
-        cdef int n_node_samples_right
+
+        print depth, " ",
+
+        if depth != 0:
+            X_current = X[sample_indices, :]
+            y_current = y[sample_indices, :]
+        else:
+            X_current = X
+            y_current = y
 
         # Count samples
-        if n_node_samples == 0:
+        if n_samples == 0:
             raise ValueError("Attempting to find a split "
-                             "with an empty sample_mask")
+                             "with no samples")
+        assert( len(sample_indices) == n_samples)
 
         # Split samples
         if depth < self.max_depth and \
-           n_node_samples >= self.min_samples_split and \
-           n_node_samples >= 2 * self.min_samples_leaf:
-            self.find_split(X_ptr, X_stride,
-                            X_argsorted_ptr, X_argsorted_stride,
-                            y_ptr, y_stride,
-                            sample_mask_ptr,
-                            n_node_samples,
-                            n_total_samples,
-                            &feature, &threshold, &best_error, &init_error)
+           n_samples >= self.min_samples_split and \
+           n_samples >= 2 * self.min_samples_leaf:
+            self.find_split(X_current, y_current, n_samples, &feature, &threshold, &best_error, &init_error)
 
         else:
             feature = -1
-            criterion.init(y_ptr, y_stride, sample_mask_ptr, n_node_samples, n_total_samples)
+            y_ptr = <DOUBLE_t*> y_current.data
+            y_stride = <int> y_current.strides[0] / <int> y_current.strides[1]
+            criterion.init(y_ptr, y_stride, n_samples)
             init_error = criterion.eval()
 
         criterion.init_value(buffer_value)
 
         # Current node is leaf
         if feature == -1:
-            self.add_leaf(parent, is_left_child, buffer_value, init_error, n_node_samples)
+            self.add_leaf(parent, is_left_child, buffer_value, init_error, n_samples)
 
         # Current node is internal node (= split node)
         else:
-            # Sample mask is too sparse?
-            if 1. * n_node_samples / n_total_samples <= self.min_density:
-                X = X[sample_mask]
-                X_argsorted = np.asfortranarray(np.argsort(X.T, axis=1).astype(np.int32).T)
-                y = y[sample_mask]
-                sample_mask = np.ones((n_node_samples, ), dtype=np.bool)
-
-                n_total_samples = n_node_samples
-
-                X_ptr = <DTYPE_t*> X.data
-                X_stride = <int> X.strides[1] / <int> X.strides[0]
-                sample_mask_ptr = <BOOL_t*> sample_mask.data
-
-                # !! No need to update the other variables
-                # X_argsorted_ptr = <int*> X_argsorted.data
-                # y_ptr = <DOUBLE_t*> y.data
-                # X_argsorted_stride = <int> X_argsorted.strides[1] / <int> X_argsorted.strides[0]
-                # y_stride = <int> y.strides[0] / <int> y.strides[1]
-
             # Split
-            X_ptr = X_ptr + feature * X_stride
-            sample_mask_left = np.zeros((n_total_samples, ), dtype=np.bool)
-            sample_mask_right = np.zeros((n_total_samples, ), dtype=np.bool)
-            n_node_samples_left = 0
-            n_node_samples_right = 0
+            n_samples_left = 0
+            n_samples_right = 0
 
-            for i from 0 <= i < n_total_samples:
-                if sample_mask_ptr[i]:
-                    if X_ptr[i] <= threshold:
-                        sample_mask_left[i] = 1
-                        n_node_samples_left += 1
-                    else:
-                        sample_mask_right[i] = 1
-                        n_node_samples_right += 1
+            tmp_left = []
+            tmp_right = []
+
+            for i from 0 <= i < n_samples:
+                index = sample_indices[i]
+                assert(index < X.shape[0])
+
+                if X[index, feature] <= threshold:
+                    tmp_left.append(index)
+                    n_samples_left += 1
+                else:
+                    tmp_right.append(index)
+                    n_samples_right += 1
+
+            sample_indices_left = np.array(tmp_left)
+            sample_indices_right = np.array(tmp_right)
 
             node_id = self.add_split_node(parent, is_left_child, feature,
                                           threshold, buffer_value, best_error,
-                                          init_error, n_node_samples)
+                                          init_error, n_samples)
+
+            X_current = None
+            y_current = None
 
             # Left child recursion
-            self.recursive_partition(X, X_argsorted, y, sample_mask_left,
-                                     n_node_samples_left, depth + 1, node_id,
+            self.recursive_partition(X, y, sample_indices_left, n_samples_left, depth + 1, node_id,
                                      True, buffer_value)
 
             # Right child recursion
-            self.recursive_partition(X, X_argsorted, y, sample_mask_right,
-                                     n_node_samples_right, depth + 1, node_id,
+            self.recursive_partition(X, y, sample_indices_right, n_samples_right, depth + 1, node_id,
                                      False, buffer_value)
 
     cdef int add_split_node(self, int parent, int is_left_child, int feature,
@@ -557,32 +538,25 @@ cdef class Tree:
 
         return node_id
 
-    cdef void find_split(self, DTYPE_t* X_ptr, int X_stride,
-                         int* X_argsorted_ptr, int X_argsorted_stride,
-                         DOUBLE_t* y_ptr, int y_stride, BOOL_t* sample_mask_ptr,
-                         int n_node_samples, int n_total_samples, int* _best_i,
+    cdef void find_split(self,
+                         np.ndarray[DTYPE_t, ndim=2, mode="fortran"] X,
+                         np.ndarray[DOUBLE_t, ndim=2, mode="c"] y,
+                         int n_samples, int* _best_i,
                          double* _best_t, double* _best_error,
                          double* _initial_error):
         """Find the best dimension and threshold that minimises the error."""
         if self.find_split_algorithm == _TREE_SPLIT_BEST:
-            self.find_best_split(X_ptr, X_stride, X_argsorted_ptr,
-                                 X_argsorted_stride, y_ptr, y_stride,
-                                 sample_mask_ptr, n_node_samples,
-                                 n_total_samples, _best_i, _best_t,
+            self.find_best_split(X, y, n_samples, _best_i, _best_t,
                                  _best_error, _initial_error)
 
         elif self.find_split_algorithm == _TREE_SPLIT_RANDOM:
-            self.find_random_split(X_ptr, X_stride, X_argsorted_ptr,
-                                   X_argsorted_stride, y_ptr, y_stride,
-                                   sample_mask_ptr, n_node_samples,
-                                   n_total_samples, _best_i, _best_t,
+            self.find_random_split(X, y, n_samples, _best_i, _best_t,
                                    _best_error, _initial_error)
 
-    cdef void find_best_split(self, DTYPE_t* X_ptr, int X_stride,
-                              int* X_argsorted_ptr, int X_argsorted_stride,
-                              DOUBLE_t* y_ptr, int y_stride,
-                              BOOL_t* sample_mask_ptr, int n_node_samples,
-                              int n_total_samples, int* _best_i,
+    cdef void find_best_split(self,
+                              np.ndarray[DTYPE_t, ndim=2, mode="fortran"] X,
+                              np.ndarray[DOUBLE_t, ndim=2, mode="c"] y,
+                              int n_samples, int* _best_i,
                               double* _best_t, double* _best_error,
                               double* _initial_error):
         """Implementation of `find_split` that looks for the best threshold."""
@@ -600,14 +574,26 @@ cdef class Tree:
         cdef double t, initial_error, error
         cdef double best_error = INFINITY, best_t = INFINITY
 
-        cdef DTYPE_t* X_i = NULL
-        cdef int* X_argsorted_i = NULL
+        cdef np.ndarray[DTYPE_t, ndim=1, mode="c"] X_copy = X[:, 0].copy()
+        cdef np.ndarray[np.int32_t, ndim=1, mode="c"] n_samples_arange = np.arange(n_samples, dtype=np.int32)
+        cdef DTYPE_t* X_copy_ptr
+
+        cdef DTYPE_t* X_ptr = <DTYPE_t*> X.data
+        cdef DOUBLE_t* y_ptr = <DOUBLE_t*> y.data
+
+        cdef int X_stride = <int> X.strides[1] / <int> X.strides[0]
+        cdef int y_stride = <int> y.strides[0] / <int> y.strides[1]
+
+        cdef np.ndarray[DTYPE_t, ndim=1, mode="c"] X_i = None
+        cdef DTYPE_t* X_i_ptr = NULL
+        cdef np.ndarray[np.int32_t, ndim=1, mode="c"] X_argsorted_i = np.zeros((n_samples,), dtype=np.int32)
+        cdef int* X_argsorted_i_ptr = NULL
         cdef DTYPE_t X_a, X_b
 
         cdef np.ndarray[np.int32_t, ndim=1, mode="c"] features = None
 
         # Compute the initial criterion value in the node
-        criterion.init(y_ptr, y_stride, sample_mask_ptr, n_node_samples, n_total_samples)
+        criterion.init(y_ptr, y_stride, n_samples)
         initial_error = criterion.eval()
 
         if initial_error == 0:  # break early if the node is pure
@@ -633,9 +619,19 @@ cdef class Tree:
         for feature_idx from 0 <= feature_idx < max_features:
             i = features[feature_idx]
 
-            # Get i-th col of X and X_sorted
-            X_i = X_ptr + X_stride * i
-            X_argsorted_i = X_argsorted_ptr + X_argsorted_stride * i
+            # Get i-th col of X and X_argsorted
+            #X_i_ptr = X_ptr + i * X_stride
+            #X_copy_ptr = <DTYPE_t*> X_copy.data
+            #memcpy(X_copy_ptr, X_i_ptr, n_samples * sizeof(DTYPE_t))
+            #X_argsorted_i_ptr = <int*> X_argsorted_i.data
+            #memcpy(X_argsorted_i_ptr, <int*> n_samples_arange.data, n_samples * sizeof(int))
+            #_argqsort(X_copy_ptr, X_argsorted_i_ptr, 0, n_samples-1)
+
+            # Get i-th col of X and X_argsorted            
+            X_i = X[:, i]
+            X_i_ptr = <DTYPE_t*> X_i.data
+            X_argsorted_i = np.argsort(X_i, kind='heapsort').astype(np.int32)
+            X_argsorted_i_ptr = <int*> X_argsorted_i.data
 
             # Reset the criterion for this feature
             criterion.reset()
@@ -643,30 +639,26 @@ cdef class Tree:
             # Index of smallest sample in X_argsorted_i that is in the sample mask
             a = 0
 
-            while sample_mask_ptr[X_argsorted_i[a]] == 0:
-                a = a + 1
-
             # Consider splits between two consecutive samples
             while True:
                 # Find the following larger sample
-                b = _smallest_sample_larger_than(a, X_i, X_argsorted_i,
-                                                 sample_mask_ptr, n_total_samples)
+                b = _smallest_sample_larger_than(a, X_i_ptr, X_argsorted_i_ptr, n_samples)
                 if b == -1:
                     break
 
                 # Better split than the best so far?
-                n_left = criterion.update(a, b, y_ptr, y_stride, X_argsorted_i, sample_mask_ptr)
+                n_left = criterion.update(a, b, y_ptr, y_stride, X_argsorted_i_ptr)
 
                 # Only consider splits that respect min_leaf
-                if n_left < min_samples_leaf or (n_node_samples - n_left) < min_samples_leaf:
+                if n_left < min_samples_leaf or (n_samples - n_left) < min_samples_leaf:
                     a = b
                     continue
 
                 error = criterion.eval()
 
                 if error < best_error:
-                    X_a = X_i[X_argsorted_i[a]]
-                    X_b = X_i[X_argsorted_i[b]]
+                    X_a = X_i_ptr[X_argsorted_i_ptr[a]]
+                    X_b = X_i_ptr[X_argsorted_i_ptr[b]]
 
                     t = X_a + (X_b - X_a) / 2.0
                     if t == X_b:
@@ -684,11 +676,10 @@ cdef class Tree:
         _best_error[0] = best_error
         _initial_error[0] = initial_error
 
-    cdef void find_random_split(self, DTYPE_t* X_ptr, int X_stride,
-                                int* X_argsorted_ptr, int X_argsorted_stride,
-                                DOUBLE_t* y_ptr, int y_stride,
-                                BOOL_t* sample_mask_ptr, int n_node_samples,
-                                int n_total_samples, int* _best_i,
+    cdef void find_random_split(self,
+                                np.ndarray[DTYPE_t, ndim=2, mode="fortran"] X,
+                                np.ndarray[DOUBLE_t, ndim=2, mode="c"] y,
+                                int n_samples, int* _best_i,
                                 double* _best_t, double* _best_error,
                                 double* _initial_error):
         """Implementation of `find_split` that looks for the best threshold
@@ -708,14 +699,26 @@ cdef class Tree:
         cdef double t, initial_error, error
         cdef double best_error = INFINITY, best_t = INFINITY
 
-        cdef DTYPE_t* X_i = NULL
-        cdef int* X_argsorted_i = NULL
+        cdef np.ndarray[DTYPE_t, ndim=1, mode="c"] X_copy = X[:, 0].copy()
+        cdef np.ndarray[np.int32_t, ndim=1, mode="c"] n_samples_arange = np.arange(n_samples, dtype=np.int32)
+        cdef DTYPE_t* X_copy_ptr
+
+        cdef DTYPE_t* X_ptr = <DTYPE_t*> X.data
+        cdef DOUBLE_t* y_ptr = <DOUBLE_t*> y.data
+
+        cdef int X_stride = <int> X.strides[1] / <int> X.strides[0]
+        cdef int y_stride = <int> y.strides[0] / <int> y.strides[1]
+
+        cdef np.ndarray[DTYPE_t, ndim=1, mode="c"] X_i = None
+        cdef DTYPE_t* X_i_ptr = NULL
+        cdef np.ndarray[np.int32_t, ndim=1, mode="c"] X_argsorted_i = np.zeros((n_samples,), dtype=np.int32)
+        cdef int* X_argsorted_i_ptr = NULL
         cdef DTYPE_t X_a, X_b
 
         cdef np.ndarray[np.int32_t, ndim=1, mode="c"] features = None
 
         # Compute the initial criterion value in the node
-        criterion.init(y_ptr, y_stride, sample_mask_ptr, n_node_samples, n_total_samples)
+        criterion.init(y_ptr, y_stride, n_samples)
         initial_error = criterion.eval()
 
         if initial_error == 0:  # break early if the node is pure
@@ -741,23 +744,29 @@ cdef class Tree:
         for feature_idx from 0 <= feature_idx < max_features:
             i = features[feature_idx]
 
-            # Get i-th col of X and X_sorted
-            X_i = X_ptr + X_stride * i
-            X_argsorted_i = X_argsorted_ptr + X_argsorted_stride * i
+            # Get i-th col of X and X_argsorted
+            #X_i_ptr = X_ptr + i * X_stride
+            #X_copy_ptr = <DTYPE_t*> X_copy.data
+            #memcpy(X_copy_ptr, X_i_ptr, n_samples * sizeof(DTYPE_t))
+            #X_argsorted_i_ptr = <int*> X_argsorted_i.data
+            #memcpy(X_argsorted_i_ptr, <int*> n_samples_arange.data, n_samples * sizeof(int))
+            #_argqsort(X_copy_ptr, X_argsorted_i_ptr, 0, n_samples-1)
+
+            # Get i-th col of X and X_argsorted            
+            X_i = X[:, i]
+            X_i_ptr = <DTYPE_t*> X_i.data
+            X_argsorted_i = np.argsort(X_i, kind='heapsort').astype(np.int32)
+            X_argsorted_i_ptr = <int*> X_argsorted_i.data
 
             # Reset the criterion for this feature
             criterion.reset()
 
             # Find min and max
             a = 0
-            while sample_mask_ptr[X_argsorted_i[a]] == 0:
-                a = a + 1
-            X_a = X_i[X_argsorted_i[a]]
+            X_a = X_i_ptr[X_argsorted_i_ptr[a]]
 
-            b = n_total_samples - 1
-            while sample_mask_ptr[X_argsorted_i[b]] == 0:
-                b = b - 1
-            X_b = X_i[X_argsorted_i[b]]
+            b = n_samples - 1
+            X_b = X_i_ptr[X_argsorted_i_ptr[b]]
 
             if b <= a or X_a == X_b:
                 continue
@@ -772,18 +781,16 @@ cdef class Tree:
             c = a + 1
 
             while True:
-                if sample_mask_ptr[X_argsorted_i[c]] != 0:
-                    # FIXME why is t cast to DTYPE_t?
-                    if X_i[X_argsorted_i[c]] > (<DTYPE_t> t) or c == b:
-                        break
+                if X_i_ptr[X_argsorted_i_ptr[c]] > (<DTYPE_t> t) or c == b:
+                    break
 
                 c += 1
 
             # Better than the best so far?
-            n_left = criterion.update(0, c, y_ptr, y_stride, X_argsorted_i, sample_mask_ptr)
+            n_left = criterion.update(0, c, y_ptr, y_stride, X_argsorted_i_ptr)
             error = criterion.eval()
 
-            if n_left < min_samples_leaf or (n_node_samples - n_left) < min_samples_leaf:
+            if n_left < min_samples_leaf or (n_samples - n_left) < min_samples_leaf:
                 continue
 
             if error < best_error:
@@ -908,8 +915,7 @@ cdef class Tree:
 cdef class Criterion:
     """Interface for splitting criteria (regression and classification)."""
 
-    cdef void init(self, DOUBLE_t* y, int y_stride, BOOL_t*
-                   sample_mask, int n_samples, int n_total_samples):
+    cdef void init(self, DOUBLE_t* y, int y_stride, int n_samples):
         """Initialise the criterion."""
         pass
 
@@ -918,7 +924,7 @@ cdef class Criterion:
         pass
 
     cdef int update(self, int a, int b, DOUBLE_t* y, int y_stride,
-                    int* X_argsorted_i, BOOL_t* sample_mask):
+                    int* X_argsorted_i):
         """Update the criteria for each value in interval [a,b) (where a and b
            are indices in `X_argsorted_i`)."""
         pass
@@ -1028,8 +1034,7 @@ cdef class ClassificationCriterion(Criterion):
     def __setstate__(self, d):
         pass
 
-    cdef void init(self, DOUBLE_t* y, int y_stride, BOOL_t *sample_mask,
-                   int n_samples, int n_total_samples):
+    cdef void init(self, DOUBLE_t* y, int y_stride, int n_samples):
         """Initialise the criterion."""
         cdef int n_outputs = self.n_outputs
         cdef int* n_classes = self.n_classes
@@ -1046,9 +1051,7 @@ cdef class ClassificationCriterion(Criterion):
             for c from 0 <= c < n_classes[k]:
                 label_count_init[k * label_count_stride + c] = 0
 
-        for j from 0 <= j < n_total_samples:
-            if sample_mask[j] == 0:
-                continue
+        for j from 0 <= j < n_samples:
 
             for k from 0 <= k < n_outputs:
                 c = <int>y[j * y_stride + k]
@@ -1079,7 +1082,7 @@ cdef class ClassificationCriterion(Criterion):
                 label_count_right[k * label_count_stride + c] = label_count_init[k * label_count_stride + c]
 
     cdef int update(self, int a, int b, DOUBLE_t* y, int y_stride,
-                    int* X_argsorted_i, BOOL_t* sample_mask):
+                    int* X_argsorted_i):
         """Update the criteria for each value in interval [a,b) (where a and b
            are indices in `X_argsorted_i`)."""
         cdef int n_outputs = self.n_outputs
@@ -1094,9 +1097,6 @@ cdef class ClassificationCriterion(Criterion):
         # post condition: all samples from [0:b) are on the left side
         for idx from a <= idx < b:
             s = X_argsorted_i[idx]
-
-            if sample_mask[s] == 0:
-                continue
 
             for k from 0 <= k < n_outputs:
                 c = <int>y[s * y_stride + k]
@@ -1345,8 +1345,7 @@ cdef class RegressionCriterion(Criterion):
     def __setstate__(self, d):
         pass
 
-    cdef void init(self, DOUBLE_t* y, int y_stride, BOOL_t* sample_mask,
-                   int n_samples, int n_total_samples):
+    cdef void init(self, DOUBLE_t* y, int y_stride, int n_samples):
         """Initialise the criterion class; assume all samples
            are in the right branch and store the mean and squared
            sum in `self.mean_init` and `self.sq_sum_init`. """
@@ -1377,10 +1376,7 @@ cdef class RegressionCriterion(Criterion):
         cdef int j = 0
         cdef DOUBLE_t y_jk = 0.0
 
-        for j from 0 <= j < n_total_samples:
-            if sample_mask[j] == 0:
-                continue
-
+        for j from 0 <= j < n_samples:
             for k from 0 <= k < n_outputs:
                 y_jk = y[j * y_stride + k]
                 sq_sum_init[k] += y_jk * y_jk
@@ -1423,8 +1419,7 @@ cdef class RegressionCriterion(Criterion):
             var_left[k] = 0.0
             var_right[k] = sq_sum_right[k] - n_samples * (mean_right[k] * mean_right[k])
 
-    cdef int update(self, int a, int b, DOUBLE_t* y, int y_stride,
-                    int* X_argsorted_i, BOOL_t* sample_mask):
+    cdef int update(self, int a, int b, DOUBLE_t* y, int y_stride, int* X_argsorted_i ):
         """Update the criteria for each value in interval [a,b) (where a and b
            are indices in `X_argsorted_i`)."""
         cdef double* mean_left = self.mean_left
@@ -1445,9 +1440,6 @@ cdef class RegressionCriterion(Criterion):
         # post condition: all samples from [0:b) are on the left side
         for idx from a <= idx < b:
             j = X_argsorted_i[idx]
-
-            if sample_mask[j] == 0:
-                continue
 
             for k from 0 <= k < n_outputs:
                 y_idx = y[j * y_stride + k]
@@ -1525,8 +1517,7 @@ cdef inline np.ndarray doublep_to_ndarray(double* data, int size):
 cdef inline int _smallest_sample_larger_than(int sample_idx,
                                              DTYPE_t* X_i,
                                              int* X_argsorted_i,
-                                             BOOL_t* sample_mask,
-                                             int n_total_samples):
+                                             int n_samples):
     """Find the largest next sample.
 
     Find the index in the `X_i` array for sample who's feature
@@ -1547,11 +1538,8 @@ cdef inline int _smallest_sample_larger_than(int sample_idx,
     if sample_idx > -1:
         threshold = X_i[X_argsorted_i[sample_idx]]
 
-    for idx from sample_idx < idx < n_total_samples:
+    for idx from sample_idx < idx < n_samples:
         j = X_argsorted_i[idx]
-
-        if sample_mask[j] == 0:
-            continue
 
         if X_i[j] > threshold + 1.e-7:
             return idx
@@ -1592,3 +1580,39 @@ def _random_sample_mask(int n_total_samples, int n_total_in_bag, random_state):
             n_bagged += 1
 
     return sample_mask.astype(np.bool)
+
+cdef void _argqsort(DTYPE_t* primary, int* secondary, int l, int u):
+    """Argsort. Sedgewick's version of Lomuto, with sentinel."""
+    cdef int i = u + 1
+    cdef int m = u + 1
+    cdef BOOL_t inner
+    cdef BOOL_t outer
+    cdef DTYPE_t temp_dtype
+    cdef int temp_int
+
+    if (l >= u):
+        return
+
+    outer = True
+    while outer:
+        inner = True
+        while inner:
+            i -= 1
+            inner = (primary[i] < primary[l])
+
+        m -= 1
+
+        # swap primary[m] and primary[i]
+        temp_dtype = primary[m]
+        primary[m] = primary[i]
+        primary[i] = temp_dtype
+
+        # swap secondary[m] and secondary[i]
+        temp_int = secondary[m]
+        secondary[m] = secondary[i]
+        secondary[i] = temp_int
+
+        outer = (i > l)
+
+    _argqsort(primary, secondary, l, m - 1);
+    _argqsort(primary, secondary, m + 1, u);